+++ app/packages/looker/src/elements/index.ts
@@ -235,8 +235,8 @@ export const getImaVidElements: GetElements<ImaVidState> = (
           { node: imavid.SeekBarElement },
           { node: imavid.SeekBarThumbElement },
           { node: imavid.PlayButtonElement },
-          { node: imavid.TimeElement },
-          video.PLAYBACK_RATE,
+          { node: imavid.FrameCountElement },
+          imavid.IMAVID_PLAYBACK_RATE,
           { node: common.PlusElement },
           { node: common.MinusElement },
           { node: common.CropToContentButtonElement },
