+++ app/packages/looker/src/lookers/imavid/constants.ts
@@ -1,4 +1,4 @@
-export const DEFAULT_FRAME_RATE = 24;
+export const DEFAULT_FRAME_RATE = 25;
 export const BUFFERS_REFRESH_TIMEOUT_YIELD = 10;
 // todo: cache by bytes and not by number of samples
 export const MAX_FRAME_SAMPLES_CACHE_SIZE = 500;
